Move proxies to their own package
