+++ src/main/java/mcjty/deepresonance/DeepResonance.java
@@ -10,6 +10,7 @@ import mcjty.deepresonance.blocks.ModBlocks;
 import mcjty.deepresonance.compat.CompatHandler;
 import mcjty.deepresonance.compat.handlers.ComputerCraftCompatHandler;
 import mcjty.deepresonance.grid.WorldGridRegistry;
+import mcjty.deepresonance.proxy.CommonProxy;
 import mcjty.gui.GuiStyle;
 import mcjty.varia.Logging;
 import net.minecraft.creativetab.CreativeTabs;
@@ -31,7 +32,7 @@ public class DeepResonance implements ModBase {
     public static final String MIN_MCJTYLIB_VER = "1.2.0";
     public static final String MIN_ELECCORE_VER = "1.4.122";
 
-    @SidedProxy(clientSide="mcjty.deepresonance.ClientProxy", serverSide="mcjty.deepresonance.ServerProxy")
+    @SidedProxy(clientSide="mcjty.deepresonance.proxy.ClientProxy", serverSide="mcjty.deepresonance.proxy.ServerProxy")
     public static CommonProxy proxy;
 
     @Mod.Instance("deepresonance")
