+++ src/main/java/mcjty/deepresonance/compat/handlers/EFluxCompatHandler.java
@@ -1,0 +1,18 @@
+package mcjty.deepresonance.compat.handlers;
+
+import elec332.core.util.AbstractCompatHandler;
+
+/**
+ * Created by Elec332 on 5-8-2015.
+ */
+public class EFluxCompatHandler extends AbstractCompatHandler.ICompatHandler {
+    @Override
+    public String getName() {
+        return "EFlux";
+    }
+
+    @Override
+    public void init() {
+
+    }
+}
